add pagination

+++ pkg/model/paginate.go
@@ -1,11 +1,10 @@
 package model
 
 type Pagination struct {
-	Page    uint64
-	PerPage uint64
-	//Count   uint64
+	Page    int32
+	PerPage int32
 }
 
-func (p *Pagination) Offset() uint64 {
+func (p *Pagination) Offset() int32 {
 	return (p.Page - 1) * p.PerPage
 }
